remove save button
